Move lock controller to scheduler crate
