--- plt/plt-lock-module/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "plt-lock-module"
-version = "0.1.0"
-edition = "2024"
-
-[dependencies]
-plt-block-state.workspace = true
-plt-scheduler-types.workspace = true
-plt-token-module.workspace = true
-
-concordium_base.workspace = true
