--- plt/plt-lock-module/src/lib.rs
@@ -1,2 +1,0 @@
-pub mod lock_controller;
-pub mod lock_controller_simple;
